feat: further breakdown
