+++ lib/index.js
@@ -119,9 +119,9 @@ const Pixel_Controllers = {
   }
 };
 
-function Pixel(opts) {
+function Pixel(opts, stripsInstance = new WeakMap()) {
   if (!(this instanceof Pixel)) {
-    return new Pixel(opts);
+    return new Pixel(opts, stripsInstance);
   }
 
   // we can assume this is set because the controller is set by the strip.
@@ -142,7 +142,7 @@ function Pixel(opts) {
     }
   });
 
-  pixels.set(this, this.initialize(opts));
+  pixels.set(this, this.initialize(opts, stripsInstance));
 }
 
 Pixel.prototype.off = Pixel.prototype.clear = function() {
@@ -194,8 +194,6 @@ Pixel.prototype.colour = Pixel.prototype.color = function(color, opts) {
     }
     pixel.color.rgb = pixelcolor.value;
 
-
-    // console.log(pixel.parent.gtable);
     color = ColorString.colorValue(pixelcolor.value, pixel.parent.gtable);
     if (sendmsg) {
       // TODO probably should be pulling the color off the obj rather than
